Added DatabaseConnector class to abstract away the SQL connection, Filled out http request bodies for evencreationcontroller  alongside partial implementations

+++ EventManagementAPI/Models/Ticket.cs
@@ -5,6 +5,6 @@ namespace EventManagementAPI.Models
 	{
         public String name { get; set; }
         public Event toEvent { get; set; }
-        public int price { get; set; }
+        public Double price { get; set; }
     }
 }
